Cache entire bucket contents in memory for a configurable duration
This commit aims to reduce the high number of class C transactions.  In place of
calling ListFileNames separately for every file, we call it once in a loop to
retrieve the entire bucket contents and cache that to memory.  The cache can be
configured with the following, with environment variables taking precedence:

Environment variables:
B2_CACHE_FILENAMES: true/false.  Enables or disables the cache.  Default: false.
B2_CACHE_FILENAMES_DURATION: time in seconds, or with unix suffix.  Specifies
cache duration.  Ex: "30", "5m20s".  Default: 0 (unlimited)

Remote config options:
cache-filenames
cache-filenames-duration

There is a limit on the size of the cache, which is currently set at 1 million
filenames.  If that limit is hit, the cache is marked as incomplete.  The
incomplete cache is still used as is, but the remote will fall back to the
default lookup method for any cache misses.  Those with such large buckets
should keep the cache disabled.